Misc cleanup (gm, ty, sislib)

+++ src/melee/gm/gmtou_1.c
@@ -69,7 +69,7 @@ static void sdata2_order(void)
 /* 4D4190 */ static s32 lbl_804D4190 = 0xFFFFFFFF;
 /* 4D4194 */ static s32 lbl_804D4194 = -1;
 
-/* 4D663C */ HSD_GObj* lbl_804D663C;
+/* 4D663C */ int lbl_804D663C;
 
 void fn_80196510(void)
 {
@@ -1222,14 +1222,14 @@ void fn_80198C60(void)
     PAD_STACK(24);
 
     td = gm_GetTournamentData();
-    td->x524[2] = HSD_SisLib_803A6754(0, (s32) lbl_804D663C);
+    td->x524[2] = HSD_SisLib_803A6754(0, lbl_804D663C);
     text = td->x524[2];
     text->font_size.x = 0.054945F;
     text->font_size.y = 0.08F;
     td->x524[2]->default_alignment = 1;
     td->x524[2]->default_kerning = 1;
 
-    td->x524[3] = HSD_SisLib_803A6754(0, (s32) lbl_804D663C);
+    td->x524[3] = HSD_SisLib_803A6754(0, lbl_804D663C);
     td->x524[3]->default_alignment = 1;
     td->x524[3]->default_kerning = 1;
     HSD_SisLib_803A6B98(td->x524[3], 320.0F, 250.0F, "    ");
